Tweak tmp_parcels creation function to avoid nested function call
Avoids calling a function for each and every record in crs_parcels.

Depends on linz-bde-schema version 1.1.0 or higher for the
availability of the `bde_get_app_specific` function taking
`crs_appellation` record.

For a crs_parcels with 4million records it was observed a time
reduction from 14 minutes to 10 minutes.

References #30