Drop ORDER BY in simplified tables creation code
ORDER should be decided by caller, there's no need to do it
at data creation time. According to chat with Jeremy Palmer
this ORDER BY was added to speed up computation of diffs
for incremental updates, but changes in that diff code around
2017 made the ORDER BY obsoleted.

This change has to be tested for performance implications.

NOTE: ORDER BY needed for nextval() consistency are retained